Refactor the structure of the unicorn
