+++ alto/unicorn/models/singleton.py
@@ -1,0 +1,8 @@
+class SingletonType(type):
+    def __call__(cls, *args, **kwargs):
+        try:
+            return cls.__instance
+        except AttributeError:
+            cls.__instance = super(SingletonType, cls).__call__(
+                *args, **kwargs)
+            return cls.__instance
